Split viewport_renderer.rs into viewport_renderer/ submodule directory
types.rs: vertex/uniform structs, constants, MeshBuffers
helpers.rs: offscreen target and blit bind group creation
pipeline.rs: ViewportRenderer struct and constructor
render.rs: update/render/blit methods
tests.rs: WGSL shader validation tests
